+++ src/lcd_control.h
@@ -25,14 +25,12 @@
 #define BIT_D7 BIT7
 
 
-void configureTIM2(float p_frequence);
 void configureLcdGPIO(void);
 void setLcdBusOutput(void);
 void setLcdBusInput(void);
 void delay(int cycles);
 unsigned char checkBusyFlag(void);
 void writeLCD(int p_package);
-void writeTime(unsigned int p_value);
 void instructLCD(int p_package);
 void configureLCD(void);
