+++ src/main.c
@@ -34,6 +34,7 @@ SOFTWARE.
 #include "lcd_control.h"
 
 /* Private macro */
+/* Private variables */
 uint8_t buffer[20];
 int ptr_write = 0;
 int ptr_read = 0;
@@ -41,7 +42,6 @@ int indx = 0;
 extern timerValue;
 unsigned int currentTime = 0;
 
-/* Private variables */
 /* Private function prototypes */
 /* Private functions */
 void UART4_IRQHandler(void){
@@ -84,7 +84,7 @@ int main(void)
   /* Infinite loop */
   while (1)
   {
-	  i++;
+    i++;
 
     if(currentTime != timerValue)
     {
