+++ src/timer_control.h
@@ -1,0 +1,18 @@
+/*
+ * timer_control.h
+ *
+ *  Created on: 15 nov. 2021
+ *      Author: Matheus
+ */
+
+#ifndef LCD_CONTROL_H_
+#define LCD_CONTROL_H_
+
+#include <macros_utiles.h>
+#include "stm32f4xx.h"
+#include "lcd_control.h"
+
+void configureTIM2(float p_frequence);
+void writeTime(unsigned int p_value);
+
+#endif /* TIMER_CONTROL_H_ */
